+++ app/src/main/AndroidManifest.xml
@@ -20,10 +20,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.OneHabitTwoHabitMyHabitYouHabit">
         <activity
-            android:name=".EditProfileActivity"
+            android:name=".activities.EditProfileActivity"
             android:exported="true" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".activities.LoginActivity"
             android:exported="true">
 
             <!-- Make the login activity the launcher activity. -->
@@ -37,12 +37,12 @@
             android:name="com.theartofdev.edmodo.cropper.CropImageActivity"
             android:theme="@style/Base.Theme.AppCompat"/> <!-- optional (needed if default theme has no action bar) -->
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:exported="true">
         </activity>
         <activity
-            android:name=".CreateHabitEvent"
-            android:parentActivityName=".MainActivity"/>
+            android:name=".activities.CreateHabitEvent"
+            android:parentActivityName=".activities.MainActivity"/>
     </application>
 
 </manifest>
