+++ app/src/main/java/com/example/ohthmhyh/CustomAdapterHF.java
@@ -2,15 +2,11 @@ package com.example.ohthmhyh;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.content.Intent;
-import android.location.Address;
-import android.text.Html;
 import android.view.GestureDetector;
 import android.view.LayoutInflater;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
@@ -19,7 +15,7 @@ import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.RecyclerView;
 
 
-import java.util.ArrayList;
+import com.example.ohthmhyh.database.HabitList;
 
 /**
  * A simple RecycleviewAdapter for the Habit list.
