+++ app/src/main/java/com/example/ohthmhyh/activities/EditProfileActivity.java
@@ -1,15 +1,9 @@
-package com.example.ohthmhyh;
+package com.example.ohthmhyh.activities;
 
-import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.app.Activity;
-import android.content.Intent;
 import android.graphics.Bitmap;
-import android.net.Uri;
 import android.os.Bundle;
-import android.provider.ContactsContract;
-import android.provider.MediaStore;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -18,7 +12,9 @@ import android.widget.Toast;
 
 //import com.github.dhaval2404.imagepicker.ImagePicker; Imported later
 
-import java.io.IOException;
+import com.example.ohthmhyh.database.DatabaseAdapter;
+import com.example.ohthmhyh.R;
+import com.example.ohthmhyh.entities.User;
 
 public class EditProfileActivity extends AppCompatActivity {
