+++ app/src/main/java/com/example/ohthmhyh/database/HabitList.java
@@ -1,4 +1,6 @@
-package com.example.ohthmhyh;
+package com.example.ohthmhyh.database;
+
+import com.example.ohthmhyh.entities.Habit;
 
 import java.util.ArrayList;
