+++ app/src/main/java/com/example/ohthmhyh/entities/Habit.java
@@ -1,4 +1,4 @@
-package com.example.ohthmhyh;
+package com.example.ohthmhyh.entities;
 
 
 import java.time.LocalDate;
