+++ app/src/main/java/com/example/ohthmhyh/entities/HabitEvent.java
@@ -1,4 +1,4 @@
-package com.example.ohthmhyh;
+package com.example.ohthmhyh.entities;
 
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
