+++ app/src/main/java/com/example/ohthmhyh/entities/User.java
@@ -1,4 +1,4 @@
-package com.example.ohthmhyh;
+package com.example.ohthmhyh.entities;
 
 import java.util.ArrayList;
