+++ app/src/main/java/com/example/ohthmhyh/fragments/FeedFragment.java
@@ -1,4 +1,4 @@
-package com.example.ohthmhyh;
+package com.example.ohthmhyh.fragments;
 
 import android.os.Bundle;
 
@@ -8,12 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.ohthmhyh.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link HabitsTodayFragment#newInstance} factory method to
+ * Use the {@link FeedFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class HabitsTodayFragment extends Fragment {
+public class FeedFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +26,7 @@ public class HabitsTodayFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public HabitsTodayFragment() {
+    public FeedFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class HabitsTodayFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment HabitsTodayFragment.
+     * @return A new instance of fragment FeedFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static HabitsTodayFragment newInstance(String param1, String param2) {
-        HabitsTodayFragment fragment = new HabitsTodayFragment();
+    public static FeedFragment newInstance(String param1, String param2) {
+        FeedFragment fragment = new FeedFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class HabitsTodayFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_habits_today, container, false);
+        return inflater.inflate(R.layout.fragment_feed, container, false);
     }
 }
