+++ app/src/main/java/com/example/ohthmhyh/fragments/UserFragment.java
@@ -1,4 +1,4 @@
-package com.example.ohthmhyh;
+package com.example.ohthmhyh.fragments;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -11,6 +11,10 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
 
+import com.example.ohthmhyh.R;
+import com.example.ohthmhyh.entities.User;
+import com.example.ohthmhyh.activities.EditProfileActivity;
+import com.example.ohthmhyh.activities.LoginActivity;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
