+++ app/src/main/res/layout/create_habit_event.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CreateHabitEvent">
+    tools:context=".activities.CreateHabitEvent">
 
 
     <ImageView
